python/ocr/requirements.txt
autopep8==1.5.7
certifi==2021.5.30
grpcio==1.38.0
grpcio-tools==1.38.0
minio==7.0.4
protobuf==3.17.3
pycodestyle==2.7.0
six==1.16.0
toml==0.10.2
urllib3==1.26.5
